Upgraded ract libs and provide an ErrorBoundary to help track down issues.
Added ErrorBoundry to help diagnose OverlayTrigger
Also, fixed check for user/password to correctly handle undefined along with blank credentials
Used the fullConnectionString when logging to the console, which should have enough info to diagnose issues.
Removed console.log(process.env) on success login, less noise.
upgrading react seems to fix some stuff.
upgraded libraries, react specifically. Removed ^ and ~ as that is lazy and garbage.
got the drop target working again after the react upgrade.
added package-lock. Not a fan of this, but it solves some issues.
If we are going to yarn .. maybe even better, but for now we have both.